backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Database (MySQL)
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.5

# Auth & Security
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.6

# Validation
pydantic==2.5.3
email-validator==2.1.0
